Feat: Split comoponent
